Add ability for updating metadata when updating an order
